recode/sandbox-1.0.0.145/examples/volcengine-provider/pyproject.toml
[project]
name = "volcengine-provider"
version = "0.1.0"
description = "Volcengine cloud provider example for agent-sandbox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-sandbox",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
agent-sandbox = { path = "../../sdk/python", editable = true }
